requirements.txt
Django==1.11.4
django-crispy-forms==1.6.1
django-environ==0.4.3
pytz==2017.2
six==1.10.0
